+++ packages/oauth/oauth-provider-ui/src/pages/account/(authenticated)/about/page.fr.tsx
