+++ packages/oauth/oauth-provider-ui/src/pages/account/(authenticated)/devices/page.tsx
