+++ packages/oauth/oauth-provider-ui/src/pages/account/(authenticated)/password/page.tsx
