--- packages/oauth/oauth-provider-ui/src/routes/account/(unauthenticated)/routes.tsx
@@ -1,3 +1,0 @@
-import { Route as ResetPasswordRoute } from './reset-password/route.tsx'
-
-export const routes = [ResetPasswordRoute]
